Consolidation + remove dependency on Scrutor
