+++ src/ForEvolve.ExceptionMapper.Core/CommonExceptions/ClientErrors/ResourceNotFoundException.cs
