+++ src/ForEvolve.ExceptionMapper.Core/IExceptionMappingBuilder.cs
