+++ src/ForEvolve.ExceptionMapper/DependencyInjection/ServiceCollectionWrapper.cs
